only clone if not git repository

+++ deploy/fix_node-sass.sh
@@ -3,13 +3,17 @@
 pushd $(pwd) &&
 
 cd node_modules/gulp-sass/node_modules &&
-rm -rf node-sass &&
 
-git clone --recursive https://github.com/sass/node-sass.git &&
-cd node-sass &&
+if [ ! -d node-sass/.git ]
+then
+  rm -rf node-sass &&
 
-npm install
-node-gyp rebuild &&
-npm install &&
+  git clone --recursive https://github.com/sass/node-sass.git &&
+  cd node-sass &&
+
+  npm install
+  node-gyp rebuild &&
+  npm install
+fi
 
 popd
